Noise drawing

+++ Keep Silence/ICreature.cs
@@ -3,7 +3,7 @@
     public interface ICreature
     {
         CreatureCommand MakeStep(Game game);
-        double GetNoiseLevel();
+        int GetNoiseLevel();
         void ActionInConflict(ICreature conflictedObject, Game game);
         string GetImageFileName();
         string GetHitImageFileName();
